Make log_processor.pl actually read all config.yaml variables.
Log_processor was only using the BASEREG var from the config.yaml
conf file. We make some small changes to make it replace whatever
defaults there are with the variables set in config.yaml. We also
add some missing options in the defaults.

Note: At present, if a default is both set via cli and config file,
the second will take precedence. This might be awkward from a certain
point of view. A future improvement might want to rectify this.